requirements.txt
pyyaml
markdown
python-markdown-math
graphviz
